Better distribution of gae dependent code
